refactor(toolbar): derive popover dismissal from one ownership table
Five hand-maintained exclusion lists decided when a popover closes, one
per popover, each spelling out the events that must not dismiss it - the
Settings list alone named forty-one of the event enum's variants. Adding
a control meant remembering which list it belonged in, and forgetting
closed the popover out from under the pointer the first time the control
was used.

Each event now declares the popovers it operates inside, once, and every
dismissal rule derives from that. A slice rather than one owner because
some controls genuinely belong to two: the shapes picker is anchored
beside the overflow flyout so their toggles spare each other, the three
overflow-anchored menus spare each other's toggles and the shared
scrollbar, and Open Configurator is reachable from both Session and
Settings. Getting that wrong is what the first attempt at this did, and
the existing dismissal tests caught it.

The neighbouring `unreachable!` is downgraded too: an item drag start
whose policy names no order group is metadata drift between two tables,
not an impossible state, so it now refuses the drag and logs instead of
taking the overlay down. A test pins the pairing it was asserting.
