+++ src/ui/toolbar/model/event_policy.rs
@@ -6,9 +6,31 @@ use crate::input::Tool;
 
 use super::super::{ToolbarEvent, ToolbarSideSection};
 
+/// The popover an event belongs to, if any.
+///
+/// Dismissal used to be five hand-maintained exclusion lists, one per
+/// popover, each spelling out the events that must *not* close it. Adding a
+/// control meant remembering which list it belonged in, and forgetting closed
+/// the popover out from under the pointer the first time the control was
+/// used. Declaring the owner once here derives every list instead.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum ToolbarPopover {
+    /// The overflow flyout and the shapes picker anchored beside it.
+    TopOverflow,
+    ShapePicker,
+    /// The three overflow-anchored menu popovers.
+    Canvas,
+    Session,
+    Settings,
+    /// The precise-entry popup.
+    PrecisionEntry,
+}
+
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct ToolbarEventPolicy {
     pub(crate) persistence: ToolbarPersistence,
+    /// The popovers this event operates inside, so it does not dismiss them.
+    pub(crate) popovers: &'static [ToolbarPopover],
     pub(crate) backend_route: ToolbarBackendRoute,
     pub(crate) pre_apply_effects: Vec<ToolbarPreApplyEffect>,
     pub(crate) tablet_thickness_sensitive: bool,
@@ -18,6 +40,7 @@ impl ToolbarEventPolicy {
     pub(crate) fn for_event(event: &ToolbarEvent) -> Self {
         Self {
             persistence: persistence_for_event(event),
+            popovers: popovers_for_event(event),
             backend_route: backend_route_for_event(event),
             pre_apply_effects: pre_apply_effects_for_event(event),
             tablet_thickness_sensitive: matches!(
@@ -204,6 +227,132 @@ pub(crate) fn action_for_clear_preset(slot: usize) -> Option<Action> {
         4 => Some(Action::ClearPreset4),
         5 => Some(Action::ClearPreset5),
         _ => None,
+    }
+}
+
+/// The popovers an event operates inside, and so must not dismiss.
+///
+/// One table replacing five per-popover exclusion lists. A control declares
+/// where it lives and every dismissal rule follows, instead of each popover
+/// separately remembering not to close on it - which is how a control added
+/// to a popover could close that popover the first time it was used.
+///
+/// A slice rather than one owner because some controls genuinely belong to
+/// two: the shapes picker is anchored beside the overflow flyout so their
+/// toggles spare each other, the three overflow-anchored menus spare each
+/// other's toggles and the shared scrollbar (switching between them is one
+/// gesture), and Open Configurator is reachable from both Session and
+/// Settings.
+pub(crate) fn popovers_for_event(event: &ToolbarEvent) -> &'static [ToolbarPopover] {
+    use ToolbarPopover as P;
+    const TOP_MENUS: &[ToolbarPopover] = &[P::Canvas, P::Session, P::Settings];
+
+    match event {
+        // The overflow flyout and the shapes picker anchored beside it.
+        ToolbarEvent::ToggleTopOverflow(_) | ToolbarEvent::ToggleShapePicker(_) => {
+            &[P::TopOverflow, P::ShapePicker]
+        }
+        // Shapes hosts its own inline options.
+        ToolbarEvent::ToggleFill(_) | ToolbarEvent::NudgePolygonSides(_) => &[P::ShapePicker],
+
+        ToolbarEvent::OpenPrecisionEntry(_)
+        | ToolbarEvent::CommitPrecisionEntry { .. }
+        | ToolbarEvent::CancelPrecisionEntry => &[P::PrecisionEntry],
+
+        // Switching between the menus is one gesture, and the switch itself
+        // is what closes the previous one; the scrollbar is shared.
+        ToolbarEvent::ToggleCanvasPopover(_)
+        | ToolbarEvent::ToggleSessionPopover(_)
+        | ToolbarEvent::ToggleSettingsPopover(_)
+        | ToolbarEvent::ScrollTopPopover(_) => TOP_MENUS,
+
+        // Reachable from both menus that offer it.
+        ToolbarEvent::OpenConfigurator => &[P::Session, P::Settings],
+
+        // Canvas hosts the board/page/zoom/advanced/step controls.
+        ToolbarEvent::BoardPrev
+        | ToolbarEvent::BoardNext
+        | ToolbarEvent::BoardNew
+        | ToolbarEvent::BoardDuplicate
+        | ToolbarEvent::BoardDelete
+        | ToolbarEvent::PagePrev
+        | ToolbarEvent::PageNext
+        | ToolbarEvent::PageNew
+        | ToolbarEvent::PageDuplicate
+        | ToolbarEvent::PageDelete
+        | ToolbarEvent::ZoomIn
+        | ToolbarEvent::ZoomOut
+        | ToolbarEvent::ResetZoom
+        | ToolbarEvent::ToggleZoomLock
+        | ToolbarEvent::UndoAll
+        | ToolbarEvent::RedoAll
+        | ToolbarEvent::UndoAllDelayed
+        | ToolbarEvent::RedoAllDelayed
+        | ToolbarEvent::ToggleFreeze
+        | ToolbarEvent::ToggleCustomSection(_)
+        | ToolbarEvent::ToggleDelaySliders(_)
+        | ToolbarEvent::SetCustomUndoSteps(_)
+        | ToolbarEvent::SetCustomRedoSteps(_)
+        | ToolbarEvent::CustomUndo
+        | ToolbarEvent::CustomRedo
+        | ToolbarEvent::SetCustomUndoDelay(_)
+        | ToolbarEvent::SetCustomRedoDelay(_)
+        | ToolbarEvent::SetUndoDelay(_)
+        | ToolbarEvent::SetRedoDelay(_) => &[P::Canvas],
+
+        // Session hosts the session controls.
+        ToolbarEvent::OpenSession
+        | ToolbarEvent::OpenRecentSession(_)
+        | ToolbarEvent::SaveSessionAs
+        | ToolbarEvent::SaveSessionAsConfirm(_)
+        | ToolbarEvent::SaveSessionAsCancel
+        | ToolbarEvent::SessionInfo
+        | ToolbarEvent::ClearSession => &[P::Session],
+
+        // Settings hosts the full Settings pane, including its customization
+        // sub-panel.
+        ToolbarEvent::SetToolbarLayoutMode(_)
+        | ToolbarEvent::ToggleContextAwareUi(_)
+        | ToolbarEvent::ToggleIconMode(_)
+        | ToolbarEvent::ToggleTextControls(_)
+        | ToolbarEvent::ToggleStatusBar(_)
+        | ToolbarEvent::SetStatusBarInteractive(_)
+        | ToolbarEvent::SetStatusBarItemVisible(_, _)
+        | ToolbarEvent::ToggleStatusBoardBadge(_)
+        | ToolbarEvent::ToggleStatusPageBadge(_)
+        | ToolbarEvent::ToggleFloatingBadgeAlways(_)
+        | ToolbarEvent::TogglePresetToasts(_)
+        | ToolbarEvent::ToggleInputHud(_)
+        | ToolbarEvent::TogglePresets(_)
+        | ToolbarEvent::ToggleActionsSection(_)
+        | ToolbarEvent::ToggleZoomActions(_)
+        | ToolbarEvent::ToggleActionsAdvanced(_)
+        | ToolbarEvent::ToggleBoardsSection(_)
+        | ToolbarEvent::TogglePagesSection(_)
+        | ToolbarEvent::ToggleStepSection(_)
+        | ToolbarEvent::SetToolbarItemCustomizationOpen(_)
+        | ToolbarEvent::SetToolbarItemCustomizationGroup(_)
+        | ToolbarEvent::SetStatusBarContentsOpen(_)
+        | ToolbarEvent::SetToolbarItemHidden(_, _)
+        | ToolbarEvent::MoveToolbarItem { .. }
+        | ToolbarEvent::StartToolbarItemDrag { .. }
+        | ToolbarEvent::DragToolbarItemOver { .. }
+        | ToolbarEvent::ResetToolbarItemOrder(_)
+        | ToolbarEvent::ResetToolbarItemHiddenOverrides
+        | ToolbarEvent::OpenCommandPalette
+        | ToolbarEvent::OpenConfigFile
+        | ToolbarEvent::OpenAbout
+        | ToolbarEvent::RequestRuntimeUiReset
+        | ToolbarEvent::ConfirmUnsupportedRuntimeUiReset
+        | ToolbarEvent::CancelUnsupportedRuntimeUiReset
+        | ToolbarEvent::RetryRuntimeUiPersistence
+        | ToolbarEvent::DiscardPendingRuntimeUiAndAdoptDisk
+        | ToolbarEvent::RequestPreserveInvalidRuntimeUiReset
+        | ToolbarEvent::ConfirmPreserveInvalidRuntimeUiReset
+        | ToolbarEvent::CancelPreserveInvalidRuntimeUiReset
+        | ToolbarEvent::CancelRuntimeUiRecovery => &[P::Settings],
+
+        _ => &[],
     }
 }
 
@@ -442,5 +591,58 @@ mod tests {
             }),
             None
         );
+    }
+}
+
+#[cfg(test)]
+mod popover_affinity_tests {
+    use super::*;
+
+    /// Every event that starts an item drag must carry the order group its
+    /// move is persisted under. The backend used to assert this with
+    /// `unreachable!`, taking the overlay down if the two ever disagreed;
+    /// it now refuses the drag, so the pairing is pinned here instead.
+    #[test]
+    fn an_item_drag_start_names_the_group_it_reorders() {
+        use crate::config::{ToolbarItemId, ToolbarItemOrderGroup};
+
+        for group in [
+            ToolbarItemOrderGroup::TopTools,
+            ToolbarItemOrderGroup::TopControls,
+            ToolbarItemOrderGroup::SideSections,
+        ] {
+            let event = ToolbarEvent::StartToolbarItemDrag {
+                group,
+                id: "top.tool.pen".parse::<ToolbarItemId>().expect("item id"),
+            };
+            assert_eq!(
+                persistence_for_event(&event),
+                ToolbarPersistence::RuntimeUi(ToolbarRuntimeUiPersistenceTarget::ItemOrder(group)),
+                "a drag start must persist under the group it reorders"
+            );
+        }
+    }
+
+    /// The controls a popover hosts keep it open; anything else closes it.
+    #[test]
+    fn a_popover_survives_its_own_controls_and_nothing_elses() {
+        use ToolbarPopover as P;
+
+        // Hosted controls.
+        assert!(popovers_for_event(&ToolbarEvent::ZoomIn).contains(&P::Canvas));
+        assert!(popovers_for_event(&ToolbarEvent::SessionInfo).contains(&P::Session));
+        assert!(popovers_for_event(&ToolbarEvent::ToggleIconMode(true)).contains(&P::Settings));
+        assert!(popovers_for_event(&ToolbarEvent::ToggleFill(true)).contains(&P::ShapePicker));
+
+        // Foreign controls close it.
+        assert!(!popovers_for_event(&ToolbarEvent::ZoomIn).contains(&P::Settings));
+        assert!(!popovers_for_event(&ToolbarEvent::SessionInfo).contains(&P::Canvas));
+        assert!(popovers_for_event(&ToolbarEvent::Undo).is_empty());
+
+        // The pairs that genuinely belong to two popovers.
+        let configurator = popovers_for_event(&ToolbarEvent::OpenConfigurator);
+        assert!(configurator.contains(&P::Session) && configurator.contains(&P::Settings));
+        let shapes = popovers_for_event(&ToolbarEvent::ToggleShapePicker(true));
+        assert!(shapes.contains(&P::ShapePicker) && shapes.contains(&P::TopOverflow));
     }
 }
